fixed http snapshot test
